+++ web/client-backplane/src/main/java/io/deephaven/javascript/proto/dhinternal/io/deephaven/proto/console_pb/CompletionContext.java
@@ -65,13 +65,13 @@ public class CompletionContext {
 
     public native String getTriggerCharacter();
 
-    public native double getTriggerKind();
+    public native int getTriggerKind();
 
     public native Uint8Array serializeBinary();
 
     public native void setTriggerCharacter(String value);
 
-    public native void setTriggerKind(double value);
+    public native void setTriggerKind(int value);
 
     public native CompletionContext.ToObjectReturnType0 toObject();
