+++ web/client-backplane/src/main/java/io/deephaven/javascript/proto/dhinternal/io/deephaven/proto/console_pb/VersionedTextDocumentIdentifier.java
@@ -66,13 +66,13 @@ public class VersionedTextDocumentIdentifier {
 
     public native String getUri();
 
-    public native double getVersion();
+    public native int getVersion();
 
     public native Uint8Array serializeBinary();
 
     public native void setUri(String value);
 
-    public native void setVersion(double value);
+    public native void setVersion(int value);
 
     public native VersionedTextDocumentIdentifier.ToObjectReturnType0 toObject();
