+++ web/client-backplane/src/main/java/io/deephaven/javascript/proto/dhinternal/io/deephaven/proto/session_pb/HandshakeRequest.java
@@ -199,7 +199,7 @@ public class HandshakeRequest {
     public static native HandshakeRequest.ToObjectReturnType toObject(
             boolean includeInstance, HandshakeRequest msg);
 
-    public native double getAuthProtocol();
+    public native int getAuthProtocol();
 
     public native HandshakeRequest.GetPayloadUnionType getPayload();
 
@@ -209,7 +209,7 @@ public class HandshakeRequest {
 
     public native Uint8Array serializeBinary();
 
-    public native void setAuthProtocol(double value);
+    public native void setAuthProtocol(int value);
 
     public native void setPayload(HandshakeRequest.SetPayloadValueUnionType value);
